postgeneration mode
In postgeneration mode, input is expected to be

  ^lemma<tags>/surface$

and will match dictionary entries as either that or simply ^surface$.

The output will be in the same form. Lemmas and tags are associated
with output forms based on position, on the expectation that rules
will generally be making small edits to surface forms rather than
adding, deleting, merging, splitting, or moving tokens as in the
typical operation of lsx-proc.